requirements.txt
pytz==2021.1
requests==2.26.0
snowflake-connector-python==2.6.2
